Moved LibCall enum to wasmer_types, and compiler is no longer dependant of wasmer_vm
