+++ lib/compiler/Cargo.toml
@@ -23,9 +23,6 @@ smallvec = "1.6"
 rkyv = { version = "0.7.20", optional = true }
 loupe = "0.1"
 
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-wasmer-vm = { path = "../vm", version = "=2.2.1" }
-
 [features]
 default = ["std", "enable-serde", "enable-rkyv"]
 # This feature is for compiler implementors, it enables using `Compiler` and
@@ -35,7 +32,7 @@ translator = ["wasmparser"]
 std = ["wasmer-types/std"]
 core = ["hashbrown", "wasmer-types/core"]
 enable-serde = ["serde", "serde_bytes", "wasmer-types/enable-serde"]
-enable-rkyv = ["rkyv", "wasmer-vm/enable-rkyv", "wasmer-types/enable-rkyv"]
+enable-rkyv = ["rkyv", "wasmer-types/enable-rkyv"]
 
 [badges]
 maintenance = { status = "experimental" }
