+++ lib/vm/src/libcalls.rs
@@ -43,12 +43,7 @@ use crate::table::{RawTableElement, TableElement};
 use crate::trap::{raise_lib_trap, Trap, TrapCode};
 use crate::vmcontext::VMContext;
 use crate::{on_host_stack, VMExternRef};
-use enum_iterator::IntoEnumIterator;
-use loupe::MemoryUsage;
-#[cfg(feature = "enable-rkyv")]
-use rkyv::{Archive, Deserialize as RkyvDeserialize, Serialize as RkyvSerialize};
-use serde::{Deserialize, Serialize};
-use std::fmt;
+pub use wasmer_types::LibCall;
 use wasmer_types::{
     DataIndex, ElemIndex, FunctionIndex, LocalMemoryIndex, LocalTableIndex, MemoryIndex,
     TableIndex, Type,
@@ -694,195 +689,39 @@ pub unsafe extern "C" fn wasmer_vm_raise_trap(trap_code: TrapCode) -> ! {
 #[no_mangle]
 pub static wasmer_vm_probestack: unsafe extern "C" fn() = PROBESTACK;
 
-/// The name of a runtime library routine.
-///
-/// This list is likely to grow over time.
-#[cfg_attr(
-    feature = "enable-rkyv",
-    derive(RkyvSerialize, RkyvDeserialize, Archive)
-)]
-#[derive(
-    Copy, Clone, Debug, PartialEq, Eq, Hash, Serialize, Deserialize, MemoryUsage, IntoEnumIterator,
-)]
-pub enum LibCall {
-    /// ceil.f32
-    CeilF32,
-
-    /// ceil.f64
-    CeilF64,
-
-    /// floor.f32
-    FloorF32,
-
-    /// floor.f64
-    FloorF64,
-
-    /// nearest.f32
-    NearestF32,
-
-    /// nearest.f64
-    NearestF64,
-
-    /// trunc.f32
-    TruncF32,
-
-    /// trunc.f64
-    TruncF64,
-
-    /// memory.size for local functions
-    Memory32Size,
-
-    /// memory.size for imported functions
-    ImportedMemory32Size,
-
-    /// table.copy
-    TableCopy,
-
-    /// table.init
-    TableInit,
-
-    /// table.fill
-    TableFill,
-
-    /// table.size for local tables
-    TableSize,
-
-    /// table.size for imported tables
-    ImportedTableSize,
-
-    /// table.get for local tables
-    TableGet,
-
-    /// table.get for imported tables
-    ImportedTableGet,
-
-    /// table.set for local tables
-    TableSet,
-
-    /// table.set for imported tables
-    ImportedTableSet,
-
-    /// table.grow for local tables
-    TableGrow,
-
-    /// table.grow for imported tables
-    ImportedTableGrow,
-
-    /// ref.func
-    FuncRef,
-
-    /// elem.drop
-    ElemDrop,
-
-    /// memory.copy for local memories
-    Memory32Copy,
-
-    /// memory.copy for imported memories
-    ImportedMemory32Copy,
-
-    /// memory.fill for local memories
-    Memory32Fill,
-
-    /// memory.fill for imported memories
-    ImportedMemory32Fill,
-
-    /// memory.init
-    Memory32Init,
-
-    /// data.drop
-    DataDrop,
-
-    /// A custom trap
-    RaiseTrap,
-
-    /// probe for stack overflow. These are emitted for functions which need
-    /// when the `enable_probestack` setting is true.
-    Probestack,
-}
-
-impl LibCall {
-    /// The function pointer to a libcall
-    pub fn function_pointer(self) -> usize {
-        match self {
-            Self::CeilF32 => wasmer_vm_f32_ceil as usize,
-            Self::CeilF64 => wasmer_vm_f64_ceil as usize,
-            Self::FloorF32 => wasmer_vm_f32_floor as usize,
-            Self::FloorF64 => wasmer_vm_f64_floor as usize,
-            Self::NearestF32 => wasmer_vm_f32_nearest as usize,
-            Self::NearestF64 => wasmer_vm_f64_nearest as usize,
-            Self::TruncF32 => wasmer_vm_f32_trunc as usize,
-            Self::TruncF64 => wasmer_vm_f64_trunc as usize,
-            Self::Memory32Size => wasmer_vm_memory32_size as usize,
-            Self::ImportedMemory32Size => wasmer_vm_imported_memory32_size as usize,
-            Self::TableCopy => wasmer_vm_table_copy as usize,
-            Self::TableInit => wasmer_vm_table_init as usize,
-            Self::TableFill => wasmer_vm_table_fill as usize,
-            Self::TableSize => wasmer_vm_table_size as usize,
-            Self::ImportedTableSize => wasmer_vm_imported_table_size as usize,
-            Self::TableGet => wasmer_vm_table_get as usize,
-            Self::ImportedTableGet => wasmer_vm_imported_table_get as usize,
-            Self::TableSet => wasmer_vm_table_set as usize,
-            Self::ImportedTableSet => wasmer_vm_imported_table_set as usize,
-            Self::TableGrow => wasmer_vm_table_grow as usize,
-            Self::ImportedTableGrow => wasmer_vm_imported_table_grow as usize,
-            Self::FuncRef => wasmer_vm_func_ref as usize,
-            Self::ElemDrop => wasmer_vm_elem_drop as usize,
-            Self::Memory32Copy => wasmer_vm_memory32_copy as usize,
-            Self::ImportedMemory32Copy => wasmer_vm_imported_memory32_copy as usize,
-            Self::Memory32Fill => wasmer_vm_memory32_fill as usize,
-            Self::ImportedMemory32Fill => wasmer_vm_memory32_fill as usize,
-            Self::Memory32Init => wasmer_vm_memory32_init as usize,
-            Self::DataDrop => wasmer_vm_data_drop as usize,
-            Self::Probestack => wasmer_vm_probestack as usize,
-            Self::RaiseTrap => wasmer_vm_raise_trap as usize,
-        }
-    }
-
-    /// Return the function name associated to the libcall.
-    pub fn to_function_name(&self) -> &str {
-        match self {
-            Self::CeilF32 => "wasmer_vm_f32_ceil",
-            Self::CeilF64 => "wasmer_vm_f64_ceil",
-            Self::FloorF32 => "wasmer_vm_f32_floor",
-            Self::FloorF64 => "wasmer_vm_f64_floor",
-            Self::NearestF32 => "wasmer_vm_f32_nearest",
-            Self::NearestF64 => "wasmer_vm_f64_nearest",
-            Self::TruncF32 => "wasmer_vm_f32_trunc",
-            Self::TruncF64 => "wasmer_vm_f64_trunc",
-            Self::Memory32Size => "wasmer_vm_memory32_size",
-            Self::ImportedMemory32Size => "wasmer_vm_imported_memory32_size",
-            Self::TableCopy => "wasmer_vm_table_copy",
-            Self::TableInit => "wasmer_vm_table_init",
-            Self::TableFill => "wasmer_vm_table_fill",
-            Self::TableSize => "wasmer_vm_table_size",
-            Self::ImportedTableSize => "wasmer_vm_imported_table_size",
-            Self::TableGet => "wasmer_vm_table_get",
-            Self::ImportedTableGet => "wasmer_vm_imported_table_get",
-            Self::TableSet => "wasmer_vm_table_set",
-            Self::ImportedTableSet => "wasmer_vm_imported_table_set",
-            Self::TableGrow => "wasmer_vm_table_grow",
-            Self::ImportedTableGrow => "wasmer_vm_imported_table_grow",
-            Self::FuncRef => "wasmer_vm_func_ref",
-            Self::ElemDrop => "wasmer_vm_elem_drop",
-            Self::Memory32Copy => "wasmer_vm_memory32_copy",
-            Self::ImportedMemory32Copy => "wasmer_vm_imported_memory32_copy",
-            Self::Memory32Fill => "wasmer_vm_memory32_fill",
-            Self::ImportedMemory32Fill => "wasmer_vm_imported_memory32_fill",
-            Self::Memory32Init => "wasmer_vm_memory32_init",
-            Self::DataDrop => "wasmer_vm_data_drop",
-            Self::RaiseTrap => "wasmer_vm_raise_trap",
-            // We have to do this because macOS requires a leading `_` and it's not
-            // a normal function, it's a static variable, so we have to do it manually.
-            #[cfg(target_vendor = "apple")]
-            Self::Probestack => "_wasmer_vm_probestack",
-            #[cfg(not(target_vendor = "apple"))]
-            Self::Probestack => "wasmer_vm_probestack",
-        }
-    }
-}
-
-impl fmt::Display for LibCall {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        fmt::Debug::fmt(self, f)
+/// The function pointer to a libcall
+pub fn function_pointer(libcall: LibCall) -> usize {
+    match libcall {
+        LibCall::CeilF32 => wasmer_vm_f32_ceil as usize,
+        LibCall::CeilF64 => wasmer_vm_f64_ceil as usize,
+        LibCall::FloorF32 => wasmer_vm_f32_floor as usize,
+        LibCall::FloorF64 => wasmer_vm_f64_floor as usize,
+        LibCall::NearestF32 => wasmer_vm_f32_nearest as usize,
+        LibCall::NearestF64 => wasmer_vm_f64_nearest as usize,
+        LibCall::TruncF32 => wasmer_vm_f32_trunc as usize,
+        LibCall::TruncF64 => wasmer_vm_f64_trunc as usize,
+        LibCall::Memory32Size => wasmer_vm_memory32_size as usize,
+        LibCall::ImportedMemory32Size => wasmer_vm_imported_memory32_size as usize,
+        LibCall::TableCopy => wasmer_vm_table_copy as usize,
+        LibCall::TableInit => wasmer_vm_table_init as usize,
+        LibCall::TableFill => wasmer_vm_table_fill as usize,
+        LibCall::TableSize => wasmer_vm_table_size as usize,
+        LibCall::ImportedTableSize => wasmer_vm_imported_table_size as usize,
+        LibCall::TableGet => wasmer_vm_table_get as usize,
+        LibCall::ImportedTableGet => wasmer_vm_imported_table_get as usize,
+        LibCall::TableSet => wasmer_vm_table_set as usize,
+        LibCall::ImportedTableSet => wasmer_vm_imported_table_set as usize,
+        LibCall::TableGrow => wasmer_vm_table_grow as usize,
+        LibCall::ImportedTableGrow => wasmer_vm_imported_table_grow as usize,
+        LibCall::FuncRef => wasmer_vm_func_ref as usize,
+        LibCall::ElemDrop => wasmer_vm_elem_drop as usize,
+        LibCall::Memory32Copy => wasmer_vm_memory32_copy as usize,
+        LibCall::ImportedMemory32Copy => wasmer_vm_imported_memory32_copy as usize,
+        LibCall::Memory32Fill => wasmer_vm_memory32_fill as usize,
+        LibCall::ImportedMemory32Fill => wasmer_vm_memory32_fill as usize,
+        LibCall::Memory32Init => wasmer_vm_memory32_init as usize,
+        LibCall::DataDrop => wasmer_vm_data_drop as usize,
+        LibCall::Probestack => wasmer_vm_probestack as usize,
+        LibCall::RaiseTrap => wasmer_vm_raise_trap as usize,
     }
 }
